Render place detail content with isHTML flag set as HTML
To accommodate raw landmark HTML in shareabouts place models, render
content with the isHTML flag set as raw HTML in the place detail template.
Assume raw HTML content is stored under the 'description' field, and
that this field is of type 'textarea', as set in the dynamic form section
of the config.

Addresses: #407